requirements.txt
discord.py>=2.5.0
python-dotenv>=1.1.0
aiosqlite>=0.21.0